Refactor, remove ingress_* prefix.
